fix(review): harden the closeout preflight and split its script
Codex review round 1 on PR 2260.

- Resolve the `codex` executable through its own symlinks and refuse a
  target inside the repository. Only the parent directory was resolved
  before, so `~/bin/codex -> <repo>/reviewer` passed the preflight and
  ran with the credentials the review needs.
- Scrub `GIT_CONFIG_COUNT` with its `GIT_CONFIG_KEY_*`/`GIT_CONFIG_VALUE_*`
  pairs and `GIT_CONFIG_PARAMETERS` from the wrapper's own Git
  environment. `codex` runs under an allowlist that drops them, so an
  injected `core.filemode=false` hid a change from both fingerprints
  while the reviewer still saw it.
- Stop when HEAD moves during base resolution. A commit landing between
  the HEAD read and the fingerprint sits in both fingerprints, so the
  pair agreed while `head_sha` named a commit `codex` never saw.
- Split the 883-line script into `closeout-review{,-exec,-git}.mjs`. Each
  is under the 600-line soft cap, the tooling arm routes all three, and
  the file-size watchlist carries no row for them.
- Name the verifier handoff in the standard and full `review-materiality`
  advice. `agent:closeout-review` is the finder pass only; card step 4
  requires the `review` skill over the same diff.
- Say in the eval that the autoreview adapter is still in the tree at
  this commit; a later change in the stack deletes it.

Refs #2239
