+++ scripts/pr/review-materiality.mjs
@@ -321,6 +321,15 @@ function classifyBySize(fileCount, lineChanges) {
   return null;
 }
 
+/**
+ * The closeout script is the finder pass only. Operating-card step 4 requires
+ * the verifier pass over the same diff, so the advice names it too: half a
+ * two-model closeout is not a closeout.
+ */
+const VERIFIER_HANDOFF =
+  "Hand that report to the review skill for the verifier pass; " +
+  "with no codex on PATH, run the review skill alone and say so in ## Validation.";
+
 function recommendedReview(tier) {
   if (tier === "trivial") {
     return [
@@ -333,12 +342,14 @@ function recommendedReview(tier) {
     return [
       "Run the quality gate from operating-card step 3.",
       "Run pnpm agent:closeout-review before pushing.",
+      VERIFIER_HANDOFF,
     ];
   }
 
   return [
     "Run the quality gate from operating-card step 3.",
     "Run pnpm agent:closeout-review before pushing.",
+    VERIFIER_HANDOFF,
     "Read any mapped checklist and audit sibling surfaces before the next push.",
   ];
 }
